fix: root document shell so Rybbit beforeInteractive runs outside RSC
