+++ src/app/layout.js
@@ -1,3 +1,16 @@
+import { Suspense } from 'react';
+import { defaultLocale } from '@/i18n/settings';
+import '@/app/globals.css';
+import { RootDocumentAsync, RootDocumentShell } from './RootDocument';
+
 export default function RootLayout({ children }) {
-  return children;
+  return (
+    <Suspense
+      fallback={
+        <RootDocumentShell locale={defaultLocale}>{children}</RootDocumentShell>
+      }
+    >
+      <RootDocumentAsync>{children}</RootDocumentAsync>
+    </Suspense>
+  );
 }
